Improve README.md
Fixed a couple typos in `README.md`, including the organization name in
the link to the new repository (as pointed out by @0xflotus).

Fixed the formatting in accordance with Google's Markdown Style Guide.

Commented out the note at the beginning of the document without deleting
it, as it is meant for the original repository owned by @zMrKrabz.